make all services into one service which can be injected with gateway class
