+++ app/services/location/jhu.py
@@ -21,22 +21,6 @@ LOGGER = logging.getLogger("services.location.jhu")
 PID = os.getpid()
 
 
-class JhuLocationService(LocationService):
-    """
-    Service for retrieving locations from Johns Hopkins CSSE (https://github.com/CSSEGISandData/COVID-19).
-    """
-    def __init__(self):
-        self.gateway = JHUGateway("https://raw.githubusercontent.com/CSSEGISandData/2019-nCoV/master/csse_covid_19_data/csse_covid_19_time_series/")
-
-    async def get_all(self):
-        # Get the locations.
-        locations = await self.gateway.get_locations()
-        return locations
-
-    async def get(self, loc_id):  # pylint: disable=arguments-differ
-        # Get location at the index equal to provided id.
-        locations = await self.get_all()
-        return locations[loc_id]
 
 
 # ---------------------------------------------------------------
