+++ app/services/location/nyt.py
@@ -16,24 +16,6 @@ from . import LocationService, LocationGateway
 LOGGER = logging.getLogger("services.location.nyt")
 
 
-class NYTLocationService(LocationService):
-    """
-    Service for retrieving locations from New York Times (https://github.com/nytimes/covid-19-data).
-    """
-
-    def __init__(self):
-        self.gateway = NYTGateway("https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv")
-
-    async def get_all(self):
-        # Get the locations.
-        locations = await self.gateway.get_locations()
-        return locations
-
-    async def get(self, loc_id):  # pylint: disable=arguments-differ
-        # Get location at the index equal to provided id.
-        locations = await self.get_all()
-        return locations[loc_id]
-
 
 # ---------------------------------------------------------------
